requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --extra=dev --output-file=./requirements-dev.txt
#
appdirs==1.4.4
    # via black
astroid==2.5
    # via pylint
attrs==22.1.0
    # via pytest
black==20.8b1
    # via google-events (setup.py)
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   gapic-generator
gapic-generator==1.4.4
    # via google-events (setup.py)
google-api-core==2.10.2
    # via gapic-generator
google-auth==2.12.0
    # via google-api-core
googleapis-common-protos[grpc]==1.56.4
    # via
    #   gapic-generator
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.4
    # via gapic-generator
grpcio==1.49.1
    # via
    #   gapic-generator
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-tools
grpcio-tools==1.49.1
    # via google-events (setup.py)
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
jinja2==3.1.2
    # via gapic-generator
lazy-object-proxy==1.7.1
    # via astroid
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via pytest
pathspec==0.10.1
    # via black
pluggy==0.13.1
    # via pytest
proto-plus==1.22.1
    # via google-events (setup.py)
protobuf==4.21.7
    # via
    #   gapic-generator
    #   google-api-core
    #   google-events (setup.py)
    #   googleapis-common-protos
    #   grpcio-tools
    #   proto-plus
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pylint==2.6.0
    # via google-events (setup.py)
pypandoc==1.9
    # via gapic-generator
pyparsing==3.0.9
    # via packaging
pytest==6.2.2
    # via google-events (setup.py)
python-dateutil==2.8.2
    # via google-events (setup.py)
pyyaml==6.0
    # via gapic-generator
regex==2022.9.13
    # via black
requests==2.28.1
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   grpcio
    #   python-dateutil
stringcase==1.2.0
    # via google-events (setup.py)
toml==0.10.2
    # via
    #   black
    #   pylint
    #   pytest
typed-ast==1.5.4
    # via black
typing-extensions==4.4.0
    # via black
urllib3==1.26.12
    # via requests
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
